requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.0,<5.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
django-celery-beat>=2.5,<3.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
Pillow>=10.0,<11.0
celery>=5.4,<6.0
redis>=5.0,<6.0
gunicorn>=22.0,<23.0
sentry-sdk>=1.35,<2.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0
coverage>=7.4,<8.0
dj-database-url>=2.1,<3.0
django-debug-toolbar>=4.3,<5.0
ipython>=8.24,<9.0
django-extensions>=3.2,<4.0
whitenoise>=6.0,<7.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
daphne>=4.0,<5.0
django-health-check>=3.17,<4.0
pybreaker>=1.0,<2.0
requests>=2.31,<3.0
